maint: rewrite example script as makefile target
While the script `authenticate' is listed as something to run to gain
access to the cluster(s) we use, it doesn't appear to do what is expected:
executing `source' commands in the current process and then exiting is
idempotent, and getting credentials for another cluster overwrites the
previous call.

It also confuses the specification of regional and zonal resources.

Instead we can provide a makefile that does the same thing but whose
effects remain.  It only requires that `NAME' be set to the name of the
target cluster.

Signed-off-by: Alex Melbourne <a.melbourne@kew.org>
Reviewed-by: Patrick Cox <p.cox2@kew.org>
Cc: Nick Black <n.black@kew.org>