maint: dump assumption of maintainer permission
Our repository holds a collection of Helm charts for use with a Kubernetes
cluster and, for us, this exists on Google Cloud Platform (GCP).  Since it
isn't possible to achieve anything without access to this system, document
this fact in the file for the maintainer.

This means our installation instructions can focus entirely on interacting
with Kubernetes and not on the bureaucracy of being granted access.

We shift the important requirement to publish work before an installation
from the bottom of general documentation to being part of these notes.

Signed-off-by: Alex Melbourne <a.melbourne@kew.org>
Reviewed-by: Patrick Cox <p.cox2@kew.org>
Cc: Nick Black <n.black@kew.org>